chore: bump to TypeScript 5.9
viem requires TypeScript ^5.0 so we are going for the latest.
Our current eslint/prettier config packages depended on ^4.0 so
we had to bump them - this means the codebase is getting reformatted
using latest rules. Quite annoying but has to happen.

There are no logic changes in this PR, just packages bumps and
formatting changes.